--- .travis.yml
@@ -1,29 +1,0 @@
-language: android
-
-env:
-    global:
-        - ADB_INSTALL_TIMEOUT=20
-
-
-android:
-    components:
-        - tools
-        - platform-tools
-        - tools
-        - build-tools-26.0.2
-        - android-26
-        - extra-google-m2repository
-        - extra-android-m2repository
-
-script:
-        - ./gradlew test
-        - ./gradlew lint
-        - ./gradlew ktlintCheck
-
-before_cache:
-    - rm -f  $HOME/.gradle/caches/modules-2/modules-2.lock
-    - rm -fr $HOME/.gradle/caches/*/plugin-resolution/
-cache:
-    directories:
-        - $HOME/.gradle/caches/
-        - $HOME/.gradle/wrapper/
