+++ app/src/androidTest/java/com/example/sun/centroiddecomposition/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.demo
+package com.example.sun.centroiddecomposition
 
 import android.support.test.InstrumentationRegistry
 import android.support.test.runner.AndroidJUnit4
@@ -6,7 +6,7 @@ import android.support.test.runner.AndroidJUnit4
 import org.junit.Test
 import org.junit.runner.RunWith
 
-import org.junit.Assert.assertEquals
+import org.junit.Assert.*
 
 /**
  * Instrumented test, which will execute on an Android device.
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getTargetContext()
-        assertEquals("com.example.demo", appContext.packageName)
+        assertEquals("com.example.sun.centroiddecomposition", appContext.packageName)
     }
 }
