+++ app/src/test/java/com/example/sun/centroiddecomposition/ExampleUnitTest.kt
@@ -1,8 +1,8 @@
-package com.example.demo
+package com.example.sun.centroiddecomposition
 
 import org.junit.Test
 
-import org.junit.Assert.assertEquals
+import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
