+++ build.gradle
@@ -1,19 +1,18 @@
-buildscript {
-  ext.kotlin_version = '1.2.0'
+// Top-level build file where you can add configuration options common to all sub-projects/modules.
 
-  repositories {
-    google()
-    jcenter()
-    mavenCentral()
-    maven {
-      url "https://plugins.gradle.org/m2/"
+buildscript {
+    ext.kotlin_version = '1.1.51'
+    repositories {
+        google()
+        jcenter()
     }
-  }
-  dependencies {
-    classpath 'com.android.tools.build:gradle:3.0.1'
-    classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-    classpath "gradle.plugin.org.jlleitschuh.gradle:ktlint-gradle:2.3.0"
-  }
+    dependencies {
+        classpath 'com.android.tools.build:gradle:3.0.1'
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+
+        // NOTE: Do not place your application dependencies here; they belong
+        // in the individual module build.gradle files
+    }
 }
 
 allprojects {
@@ -23,12 +22,6 @@ allprojects {
     }
 }
 
-apply plugin: "org.jlleitschuh.gradle.ktlint"
-
-repositories {
-  mavenCentral()
-}
-
-ktlint {
-  ignoreFailures = false
+task clean(type: Delete) {
+    delete rootProject.buildDir
 }
