--- lib/build.gradle
@@ -1,19 +1,0 @@
-buildscript {
-  repositories {
-    jcenter()
-  }
-  dependencies {
-    classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:${rootProject.kotlin_version}"
-  }
-}
-
-apply plugin: 'kotlin'
-
-repositories {
-  mavenCentral()
-}
-
-dependencies {
-  compile "org.jetbrains.kotlin:kotlin-stdlib-jre7:${rootProject.kotlin_version}"
-  testCompile group: 'junit', name: 'junit', version: '4.12'
-}
