+++ settings.gradle
@@ -1,4 +1,1 @@
 include ':app'
-include ':lib'
-
-
